place all compiled assets in build folder
