+++ src/assets/favicon.ico
