+++ src/assets/scss/terminal.scss
