+++ src/buffer.ts
@@ -3,13 +3,13 @@ import { createInterface } from 'readline';
 ask('Enter your username');
 
 function ask(question: string): Promise<string> {
-  const r = createInterface({
+  const rlp = createInterface({
     input: process.stdin,
     output: process.stdout,
   });
   return new Promise(resolve => {
-    r.question(`${question}: `, answer => {
-      r.close();
+    rlp.question(`${question}: `, answer => {
+      rlp.close();
       resolve(answer);
     });
   });
