--- src/client/client/options.ts
@@ -1,12 +1,0 @@
-import { isUndefined } from '../../web_modules/lodash.js';
-
-export function loadOptions(): object {
-  const defaultOptions = { fontSize: 14 };
-  try {
-    return isUndefined(localStorage.options)
-      ? defaultOptions
-      : JSON.parse(localStorage.options);
-  } catch {
-    return defaultOptions;
-  }
-}
