+++ src/client/dev.ts
@@ -1,0 +1,5 @@
+caches.keys().then(cacheNames => {
+  cacheNames.forEach(cacheName => {
+    caches.delete(cacheName);
+  });
+});
