+++ src/client/shared/elements.ts
@@ -1,2 +1,5 @@
 export const overlay = document.getElementById('overlay');
 export const terminal = document.getElementById('terminal');
+export const editor = document.querySelector(
+  '#options .editor',
+) as HTMLInputElement;
